requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
decorator==5.1.1
font-fredoka-one==0.0.4
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
font-source-serif-pro==0.0.1
future==0.18.2
geocoder==1.38.1
idna==3.3
inky==1.3.1
numpy==1.22.2
Pillow==9.0.1
ratelim==0.1.6
requests==2.27.1
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.1
soupsieve==2.3.1
spidev==3.5
urllib3==1.26.8
